refactor(deploy): read project-config.sh
Kit designers can now specify the repository of the project to
capture, and the branches and commits within that repository to
capture using the following variables.

    SOURCE_REPO_CLONE_URL
    SOURCE_REPO_BRANCHES

The second is a bash array of strings. Each string has the form
"branch:commit" specifying the branch name to capture and the commit
it should be restored to.