refactor: add command/release.bash
Adding command/release.bash which will tag multi-platform Docker
images based on the image tagged with `main`. This should be automatic
but the automated release pipeline has been failing. Theoretically
command/release.bash should not be needed. But if the release job
fails... well...

Assuming that the release pipeline (that runs when merging into main)
has successfully created an image tagged with ":main" then
command/release.bash can be used as follows. Assuming the new release
version is 2.0.0

```bash
./commands/release.bash 2.0.0
```

Note that the person running this must have credentials sufficient
for reading and writing the project's container registry on GitLab.