feat!: leverage GitPod to run KitClients
This update leverages GitPod to provide students
with personal, remote machines to run and interact
with KitClients. This removes the need for students
to install Docker Desktop, or for faculty to convince
their IT staff to install Docker Desktop on lab
computers. Now, students and faculty only need a
free GitPod account to launch, use, and manage
KitClients.

BREAKING CHANGES:

Activities rewritten to use the new KitClients based on GitPod.
Students now need a GitPod account.

---

Closes #106
Closes #103

Co-authored-by: Grant Braught <braughtg@gmail.com>
Co-authored-by: Stoney Jackson <dr.stoney@gmail.com>
Co-authored-by: Cam Macdonell <macdonellc4@macewan.ca>
Co-authored-by: Karl R. Wurst <karl@w-sts.com>