feat: parameterize deploy with REPO_NAME
This update allows instructors to specify the name
of the new repository. This is important since each
deployment can only support 32 students. To support
more, the instructor needs to be able to create more
than one deployment, and likely in the same namespace.
So they need to be able to give different names.

---

Closes #113